[consolidate-create-constants] cleanups and creates new jasmine_common.py file
